Add files via upload
Hi

// I will refer to Overwacth League Points as LP //
// The promgram hase been made for 1080p monitors, if yours isn't you will have to adjust the pixel position by yourself, it is a very simple process so you shouldn't have any problem here //

Requisites:

Python: I personally dont know with wich version it doesnt work all I can tell its that I'm using the lastest current  version of it wich is 3.7. 6

Libraries: 
     -    datetime
     -    time
     -    schedule
     -    webbrowser
     -    win32api and win32con
     

Program intended use:
The purpouse of the program is to check up every 60 if theres a programed overwacth match, wich time we must introduce first, and then open autamticaly the streaming so you can obtain the points. Cleraly this is neant to be running while we are NOT using the pc, but fot turning on before going to do something else, that wolundt let us be there for entering the streaming, like job, school or sleeping.

Instruction:

- Turn on

- Closing: For closing we must hit ctrl+shift+esc to open task manager and end the python process, if theres any problem don't worry the bot is programed to turn of once thres has past 9 hours withouth any  matching time


GENERAL DESCRIPTION:
This is a program ment to farm Overwatch LP, it's a very simple code, in realty its nothing more of a code that checks the current time of the computer and checks if it's one that exist in a given array where you have to manually introduce the day month of the matches (I have already introduced the play off's), once the time given matches with the actual one the program will automaticaly open a tab of tyour default browser on youtube and will click in the specific place where youtube displays live stramings (of course its not preapred to watch them where thers not live direct because then we wont get any LP), dont forget that you will need to link BLIZZARD and Youtube acounts wich is an easy external process you can look up for.